Removed Capture + Move classes and Board __deepcopy__ magic method.
Valid Capture and Move logic should be handled by the Draughts class which runs the rulesets -> only need to change Draughts class to change rulesets.
Capture and Move then become a start and end position only so are indistinguishable from each other and consist of two attributes which just isn't worth creating a class for.